Mergecode

+++ src/main/java/ClothesShop/Entity/Users.java
@@ -1,7 +1,7 @@
 package ClothesShop.Entity;
-
+import org.springframework.web.multipart.MultipartFile;
 public class Users {
-	private long id_kh;
+	private int id_kh;
 	private String ten_kh;
 	private String pass;
 	private String email_kh;
@@ -10,7 +10,7 @@ public class Users {
 	public long getId_kh() {
 		return id_kh;
 	}
-	public void setId_kh(long id_kh) {
+	public void setId_kh(int id_kh) {
 		this.id_kh = id_kh;
 	}
 	public String getTen_kh() {
